Rename `e!` to `cl_error!` and move it to `small_net::error`
